refactor: moved emission_data to engine_data folder
